+++ app/users/dashboard/finance/budgets/[id]/edit/page.tsx
@@ -2,24 +2,24 @@
 
 import { useState, use } from 'react';
 import { notFound } from 'next/navigation';
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/shadcn/button';
 import {
   Card,
   CardContent,
   CardDescription,
   CardHeader,
   CardTitle,
-} from '@/components/ui/card';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
+} from '@/components/shadcn/card';
+import { Input } from '@/components/shadcn/input';
+import { Label } from '@/components/shadcn/label';
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from '@/components/ui/select';
-import { Textarea } from '@/components/ui/textarea';
+} from '@/components/shadcn/select';
+import { Textarea } from '@/components/shadcn/textarea';
 import {
   Table,
   TableBody,
@@ -27,7 +27,7 @@ import {
   TableHead,
   TableHeader,
   TableRow,
-} from '@/components/ui/table';
+} from '@/components/shadcn/table';
 // Separator not used in this file
 import { Save, X, Plus, Trash2 } from 'lucide-react';
 import { toast } from '@/lib/styles/toast-styles';
